split up db file so we can share session across files
