+++ transformerlab/db/db.py
@@ -1,14 +1,11 @@
 import json
-import os
 import sqlite3
 
-import aiosqlite
 from sqlalchemy import select, delete, text, update
 from sqlalchemy.dialects.sqlite import insert  # Correct import for SQLite upsert
 
 # from sqlalchemy import create_engine
-from sqlalchemy.ext.asyncio import create_async_engine, AsyncSession
-from sqlalchemy.orm import sessionmaker
+from sqlalchemy.ext.asyncio import AsyncSession
 
 
 # Make sure SQLAlchemy is installed using pip install sqlalchemy[asyncio] as
@@ -19,20 +16,11 @@ from typing import AsyncGenerator
 from fastapi import Depends
 from fastapi_users.db import SQLAlchemyUserDatabase
 
-from transformerlab.shared import dirs
 from transformerlab.shared.models import models  # noqa: F401
 from transformerlab.shared.models.models import Config, Plugin
 
-db = None
-DATABASE_FILE_NAME = f"{dirs.WORKSPACE_DIR}/llmlab.sqlite3"
-DATABASE_URL = f"sqlite+aiosqlite:///{DATABASE_FILE_NAME}"
-
-# Create SQLAlchemy engines
-# engine = create_engine(DATABASE_URL, echo=True)
-async_engine = create_async_engine(DATABASE_URL, echo=False)
-
-# Create a configured "Session" class
-async_session = sessionmaker(async_engine, expire_on_commit=False, class_=AsyncSession)
+from transformerlab.db.constants import DATABASE_FILE_NAME
+from transformerlab.db.session import async_session
 
 
 def unconverted(func):
@@ -44,157 +32,12 @@ def unconverted(func):
 
 
 @unconverted
-async def migrate_workflows_non_preserving():
-    """
-    Migration function that renames workflows table as backup and creates new table
-    based on current schema definition if experiment_id is not INTEGER type or config is not JSON type
-    """
-
-    try:
-        # Check if workflows table exists
-        cursor = await db.execute("SELECT name FROM sqlite_master WHERE type='table' AND name='workflows'")
-        table_exists = await cursor.fetchone()
-        await cursor.close()
-
-        if not table_exists:
-            print("Workflows table does not exist. Skipping non-preserving migration.")
-            return
-
-        # Check column types in the current workflows table
-        cursor = await db.execute("PRAGMA table_info(workflows)")
-        columns_info = await cursor.fetchall()
-        await cursor.close()
-
-        experiment_id_type = None
-        config_type = None
-
-        for column in columns_info:
-            column_name = column[1]
-            column_type = column[2].upper()
-
-            if column_name == "experiment_id":
-                experiment_id_type = column_type
-            elif column_name == "config":
-                config_type = column_type
-
-        # Check if migration is needed based on column types
-        needs_migration = False
-        migration_reasons = []
-
-        if experiment_id_type and experiment_id_type != "INTEGER":
-            needs_migration = True
-            migration_reasons.append(f"experiment_id column type is {experiment_id_type}, expected INTEGER")
-
-        # SQLAlchemy JSON type maps to TEXT in SQLite, so we accept both
-        if config_type and config_type not in ["JSON", "TEXT"]:
-            needs_migration = True
-            migration_reasons.append(
-                f"config column type is {config_type}, expected JSON/TEXT (SQLAlchemy creates JSON as TEXT in SQLite)"
-            )
-
-        if not needs_migration:
-            # print("Column types are correct. No migration needed.")
-            return
-
-        print("Migration needed due to:")
-        for reason in migration_reasons:
-            print(f"  - {reason}")
-
-        # Check if backup table already exists and drop it
-        cursor = await db.execute("SELECT name FROM sqlite_master WHERE type='table' AND name='workflows_backup'")
-        backup_exists = await cursor.fetchone()
-        await cursor.close()
-
-        if backup_exists:
-            await db.execute("DROP TABLE workflows_backup")
-
-        # Rename current table as backup
-        await db.execute("ALTER TABLE workflows RENAME TO workflows_backup")
-
-        # Create new workflows table using SQLAlchemy schema
-        async with async_engine.begin() as conn:
-            await conn.run_sync(models.Base.metadata.create_all)
-
-        await db.commit()
-        print("Successfully created new workflows table with correct schema. Old table saved as workflows_backup.")
-
-    except Exception as e:
-        print(f"Failed to perform non-preserving migration: {e}")
-        raise e
-
-
-@unconverted
-async def init():
-    """
-    Create the database, tables, and workspace folder if they don't exist.
-    """
-    global db
-    os.makedirs(os.path.dirname(DATABASE_FILE_NAME), exist_ok=True)
-    db = await aiosqlite.connect(DATABASE_FILE_NAME)
-    await db.execute("PRAGMA journal_mode=WAL")
-    await db.execute("PRAGMA synchronous=normal")
-    await db.execute("PRAGMA busy_timeout = 30000")
-
-    # Create the tables if they don't exist
-    async with async_engine.begin() as conn:
-        await conn.run_sync(models.Base.metadata.create_all)
-
-    # Check if experiment_id column exists in workflow_runs table
-    cursor = await db.execute("PRAGMA table_info(workflow_runs)")
-    columns = await cursor.fetchall()
-    has_experiment_id = any(column[1] == "experiment_id" for column in columns)
-
-    if not has_experiment_id:
-        # Add experiment_id column
-        await db.execute("ALTER TABLE workflow_runs ADD COLUMN experiment_id INTEGER")
-
-        # Update existing workflow runs with experiment_id from their workflows
-        await db.execute("""
-            UPDATE workflow_runs 
-            SET experiment_id = (
-                SELECT experiment_id 
-                FROM workflows 
-                WHERE workflows.id = workflow_runs.workflow_id
-            )
-        """)
-        await db.commit()
-
-    print("✅ Database initialized")
-
-    print("✅ SEED DATA")
-    async with async_session() as session:
-        for name in ["alpha", "beta", "gamma"]:
-            # Check if experiment already exists
-            exists = await session.execute(select(models.Experiment).where(models.Experiment.name == name))
-            if not exists.scalar_one_or_none():
-                session.add(models.Experiment(name=name, config={}))
-        await session.commit()
-
-    # On startup, look for any jobs that are in the RUNNING state and set them to CANCELLED instead:
-    # This is to handle the case where the server is restarted while a job is running.
-    await job_cancel_in_progress_jobs()
-    # Run migrations
-    await migrate_workflows_non_preserving()
-    # await init_sql_model()
-
-    return
-
-
-@unconverted
 def get_sync_db_connection():
-    global DATABASE_FILE_NAME
     db_sync = sqlite3.connect(DATABASE_FILE_NAME, isolation_level=None)
     db_sync.execute("PRAGMA journal_mode=WAL")
     db_sync.execute("PRAGMA synchronous=normal")
     db_sync.execute("PRAGMA busy_timeout = 30000")
     return db_sync
-
-
-async def close():
-    await db.close()
-    await async_engine.dispose()
-    print("✅ Database closed")
-    return
 
 
 ###############################################
@@ -210,15 +53,6 @@ async def get_async_session() -> AsyncGenerator[AsyncSession, None]:
 
 async def get_user_db(session: AsyncSession = Depends(get_async_session)):
     yield SQLAlchemyUserDatabase(session, models.User)
-
-
-# async def init_sql_model():
-#     """
-#     Initialize the database using SQLModel.
-#     """
-#     async with async_engine.begin() as conn:
-#         await conn.run_sync(SQLModel.metadata.create_all)
-#     print("✅ SQLModel Database initialized")
 
 
 ###############
@@ -687,13 +521,6 @@ async def job_delete(job_id):
     async with async_session() as session:
         # Instead of deleting, set status to 'DELETED' for the job
         await session.execute(update(models.Job).where(models.Job.id == job_id).values(status="DELETED"))
-        await session.commit()
-    return
-
-
-async def job_cancel_in_progress_jobs():
-    async with async_session() as session:
-        await session.execute(update(models.Job).where(models.Job.status == "RUNNING").values(status="CANCELLED"))
         await session.commit()
     return
